requirements.txt
clams-python==0.5.1
mmif-python==0.4.6
lapps==0.0.2
torch==1.13.1
numpy==1.23.1
matplotlib==3.6.2
tqdm==4.64.1
scipy==1.9.3
bcolz-zipline==1.2.6
easydict==1.10
opencv_python-headless==4.7.0.68
Pillow==9.4.0
mxnet==1.9.1
scikit_learn==1.2.0
tensorboardX==2.5.1
torchvision==0.14.1
